round_2/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
pydantic==2.5.3
python-dateutil==2.8.2
